chore: update types and general cleanup
