+++ src/types/openaiResponse.ts
@@ -1,4 +1,10 @@
-import { spotifyFeatures } from "./spotifyQuery.js";
+export interface spotifyFeatures {
+    valence: number, 
+    energy: number,
+    danceability: number,
+    acousticness: number,
+    tempo: number
+}
 
 export interface openaiResponse {
     mood: string,
