+++ src/types/spotifyResponse.ts
@@ -1,4 +1,5 @@
 export interface track {
+    id: string,
     title: string,
     artist: string,
     album: string,
