+++ src/types/userInput.ts
@@ -1,5 +1,5 @@
-import { openaiQuery } from "./openaiQuery.js"
+import { openaiQuery } from "./openaiQuery"
 
 export interface userInput extends openaiQuery {
-    date: Date
+    date: string
 };
